+++ VirtGabbai/DataCache/DataCache.csproj
@@ -74,15 +74,7 @@
     <Compile Include="Models\User.cs" />
     <Compile Include="Models\Yahrtzieht.cs" />
     <Compile Include="Models\ZeraLeviContext.cs" />
-    <Compile Include="Partials\Account.cs" />
-    <Compile Include="Partials\Donation.cs" />
-    <Compile Include="Partials\Person.cs" />
-    <Compile Include="Partials\PhoneNumber.cs" />
-    <Compile Include="Partials\PhoneType.cs" />
-    <Compile Include="Partials\Privilege.cs" />
-    <Compile Include="Partials\PrivilegesGroup.cs" />
-    <Compile Include="Partials\StreetAddress.cs" />
-    <Compile Include="Partials\User.cs" />
+    <Compile Include="Models\StreetAddress.cs" />
     <Compile Include="Partials\Yahrtzieht.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
