+++ VirtGabbai/DataCache/Models/Donation.cs
@@ -16,5 +16,35 @@ namespace DataCache.Models
         public virtual Account Accounts { get; set; }
 
         public static Donation Createt_donations(int _Id, int accountNumber, string reason, double amount, DateTime donationDate, bool v) => new Donation { ID = _Id, AccountID = accountNumber, Reason = reason, Amount = amount, DonationDate = donationDate, Paid = v};
+
+        public override string ToString()
+        {
+            string returnString = "Donated for: \"" + this.Reason +
+                                  "\" Amount donated: \"" + this.Amount +
+                                  "\" Date donated: \"" + this.DonationDate.ToString("dd/MM/yyyy") + "\"";
+            if (this.Comments != string.Empty)
+            {
+                returnString += " Comments: \"" + this.Comments + "\"";
+
+            }
+            return returnString;
+        }
+
+        public override bool Equals(object obj)
+        {
+            Donation donationToCompare = (Donation)obj;
+
+            if ((donationToCompare.Comments == null) && (this.Comments != null))
+            {
+                return false;
+            }
+            return ((this.ID == donationToCompare.ID) &&
+                    (this.Amount == donationToCompare.Amount) &&
+                    (this.Comments == donationToCompare.Comments) &&
+                    (this.DonationDate == donationToCompare.DonationDate) &&
+                    (this.Reason == donationToCompare.Reason));
+        }
+
+        public override int GetHashCode() => base.GetHashCode();
     }
 }
