+++ VirtGabbai/DataCache/Models/Privilege.cs
@@ -10,5 +10,16 @@ namespace DataCache.Models
         public virtual ICollection<PrivilegesGroup> PrivilegesGroup { get; set; } = new List<PrivilegesGroup>();
 
         public static Privilege Createt_privileges(int privilegeIndex) => new Privilege { ID = privilegeIndex };
+
+        public override bool Equals(object obj)
+        {
+            Privilege comparedPrivilege = (Privilege)obj;
+            return ((ID == comparedPrivilege.ID) &&
+                    (Name == comparedPrivilege.Name));
+        }
+
+        public override int GetHashCode() => base.GetHashCode();
+
+        public override string ToString() => Name;
     }
 }
