+++ VirtGabbai/DataCache/Models/Yahrtzieht.cs
@@ -13,5 +13,20 @@ namespace DataCache.Models
         public virtual Person Person { get; set; }
 
         public static Yahrtzieht Createt_yahrtziehts(int _Id, int personId, DateTime date, string name) => new Yahrtzieht { ID = _Id, PersonID = personId, Date = date, Name = name };
+
+        public override bool Equals(object yahrComparingObj)
+        {
+            Yahrtzieht yahrComparing = (Yahrtzieht)yahrComparingObj;
+            return ((this.ID == yahrComparing.ID) &&
+                    (this.Date.Date == yahrComparing.Date.Date) &&
+                    (this.Name == yahrComparing.Name) &&
+                    (this.Relation == yahrComparing.Relation));
+        }
+
+        public override string ToString() => "Deceased's Name:\"" + this.Name + "\" " +
+            "Date:\"" + this.Date.Date.ToString("dd/MM/yyyy") + "\" " +
+            "Relation:\"" + this.Relation + "\"";
+
+        public override int GetHashCode() => base.GetHashCode();
     }
 }
