--- VirtGabbai/DataCache/Partials/Privilege.cs
@@ -1,20 +1,0 @@
-﻿namespace DataCache.Models
-{
-    public partial class Privilege
-    {
-        #region Object Methods
-
-        public override bool Equals(object obj)
-        {
-            Privilege comparedPrivilege = (Privilege)obj;
-            return ((ID == comparedPrivilege.ID) &&
-                    (Name == comparedPrivilege.Name));
-        }
-
-        public override int GetHashCode() => base.GetHashCode();
-
-        public override string ToString() => Name;
-
-        #endregion
-    }
-}
